refactor: simplify completion package
Replace custom bash-to-zsh sed conversion with cobra's native
GenZshCompletion(). Merge two files into one.
